fix: address Studio hardening review
